fine-tune/pyproject.toml
[project]
name = "fine-tune"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Paulescu", email = "plabartabajo@gmail.com" } 
]
requires-python = ">=3.12"
dependencies = [
    "chess>=1.11.2",
    "datasets>=4.0.0",
    "fire>=0.7.1",
    "huggingface-hub>=0.34.4",
    "jinja2>=3.1.6",
    "leap-bundle>=0.3.0",
    "modal>=1.1.3",
    "peft>=0.17.1",
    "pydantic-settings>=2.10.1",
    "python-chess>=1.999",
    "tqdm>=4.67.1",
    "wandb>=0.21.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "jupyter>=1.1.1",
    "ruff>=0.12.11",
]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = ["F821"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
